AclUtil 버그 수정 및 솔루션 CRUD완성

+++ teamKMS/src/main/java/com/devworker/kms/service/solution/SolutionService.java
@@ -1,13 +1,16 @@
 package com.devworker.kms.service.solution;
 
 import com.devworker.kms.component.BoardComponent;
+import com.devworker.kms.dic.PermissionType;
 import com.devworker.kms.dto.common.BoardDetailDto;
 import com.devworker.kms.dto.common.BoardDto;
 import com.devworker.kms.dto.solution.SolutionDto;
 import com.devworker.kms.entity.solution.SolutionDao;
+import com.devworker.kms.exception.NotAllowException;
 import com.devworker.kms.exception.NotExistException;
 import com.devworker.kms.repo.solution.SolutionRepo;
 import com.devworker.kms.repo.solution.SolutionRepoImpl;
+import com.devworker.kms.util.AclUtil;
 import com.devworker.kms.util.CommonUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -38,6 +41,8 @@ public class SolutionService {
 	}
 	
 	public long registerSolution(SolutionDto solutionDto) {
+		if(!AclUtil.checkPermission(PermissionType.CREATESOL))
+			throw new NotAllowException("You Have not Create Solution Board Permission");
 		solutionDto.getBoardDetailDto().setRegDate(LocalDateTime.now());
 		solutionDto.getBoardDetailDto().setUpdDate(LocalDateTime.now());
 		solutionDto.getBoardDetailDto().setUserId(CommonUtil.getCurrentUser());
@@ -47,19 +52,28 @@ public class SolutionService {
 	}
 		
 	public void editSolution(SolutionDto solutionDto) {
-		Optional<SolutionDao> solutionDao = solutionRepo.findById(solutionDto.getBoardId());
-		solutionDao.orElseThrow(() -> new NotExistException("Solution Board Not Found"));
-		solutionDto.getBoardDetailDto().setUpdDate(LocalDateTime.now());
+		BoardDetailDto dto = getSolutionById(solutionDto.getBoardId());
+		if(!AclUtil.checkPermission(dto.getUserId() , PermissionType.MODIFYSOL))
+			throw new NotAllowException("You Have not Modify Solution Board Permission");
+		dto.setUpdDate(LocalDateTime.now());
+		dto.setContents(solutionDto.getBoardDetailDto().getContents());
 		boardComponent.edit(solutionDto.getBoardDetailDto());
 	}
 
 	public void deleteSolution(long id) {
+		BoardDetailDto dto = getSolutionById(id);
+		if(!AclUtil.checkPermission(dto.getUserId() , PermissionType.DELETESOL))
+			throw new NotAllowException("You Have not Delete Solution Board Permission");
 		solutionRepo.deleteById(id);
 	}
 
 	public BoardDetailDto getSolutionById(Long id) {
 		Optional<SolutionDao> solutionDao = solutionRepo.findById(id);
 		solutionDao.orElseThrow(() -> new NotExistException("Solution Board Not Found"));
-		return boardComponent.getBoard(id);
+		BoardDetailDto dto = boardComponent.getBoard(id);
+		dto.setReadOnly(!AclUtil.checkPermission(dto.getUserId() , PermissionType.MODIFYSOL));
+		dto.setHits(dto.getHits() +1);
+		boardComponent.edit(dto);
+		return dto;
 	}
 }
